my-first-project/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["haruiz <henryruiz22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = ">=1.22,<=1.24.3"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.2"
seaborn = "^0.12.2"
pandas = "^2.0.3"
tensorflow-datasets = "^4.9.2"
fastapi = "^0.101.1"
uvicorn = "^0.23.2"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.0.5"
black = "^23.7.0"

[tool.poetry.group.tf_mac]
optional = true

[tool.poetry.group.tf_mac.dependencies]
tensorflow-macos = "^2.13.0"
tensorflow-metal = "^1.0.1"

[tool.poetry.group.tf_win]
optional = true

[tool.poetry.group.tf_win.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
